refactor(ui): start sync-dialogs.tsx split — move request/result types to sync-dialogs/types.ts
First slice of the 460-LOC sync-dialogs.tsx decomposition. Hoist the
shared request/result type definitions (DialogTone, the three
*DialogRequest variants, DuplicatePersonActorOption,
DuplicatePersonDialogResult, SyncDialogRequest, SyncDialogResult)
into a dedicated types module. sync-dialogs.tsx re-exports
DuplicatePersonActorOption since team-sync.ts imports it for its own
API.
